requirements-dev.txt
# Dependências do projeto G-Click
requests>=2.25.0
PyYAML>=6.0
python-dotenv>=0.19.0
backoff>=1.11.1
python-dateutil>=2.8.2

# Dependências do Azure Functions e Bot Framework
azure-functions>=1.17.0
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botframework-connector>=4.14.0
aiohttp>=3.8.1

# Logging e utilitários
python-json-logger>=2.0.0

# Dependências de desenvolvimento e testes
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
